Remove the libuiohook logging callback
libuiohook's logging expects a callback with varargs and there is no
easy way to support it with C#. Further, the message is formatted using
the printf format specifiers, and formatting those will need to be
implemented manually. Maybe, in a future version the logging support
will be added, but for now it will be skipped.