Remove NuGet package dependencies
The Nullable package is not needed after all.

As for hash codes, it's better to implement them manually for platforms
that don't have the HashCode struct than add a dependency for them.

Also, add .NET Core 3.1 as a target platform since it has more stuff
than .NET Standard 2.0 and is still supported.