Add reactive global hooks
Add global hooks which expose their events as observables. This is a
much better solution in my opinion, and should be used whenever
possible. The default reactive implementation doesn't need to think
where the events are handled - the clients are free to do that
themselves. I've also added an adapter of IGlobalHook to
IReactiveGlobalHook though I don't think it'll be used much.

There is a lot of duplicate code between GlobalHookBase and
SimpleReactiveGlobalHook. This is by design - I don't want to extract
common functionality since it's similar, but not quite the same.