Refactor the global hooks
Remove the Start method from global hooks, and instead add the Run and
RunAsync methods. This will give users the freedom to choose whether the
hook should run in a blocking way or not. This will also enable users to
use the global hook classes on macOS.

Also, throw an exception if Run/RunAsync is called and the hook is
already running, and wrap unexpected exceptions in HookException when
starting the hook.