Add low-level functionality providers
Add interfaces which provide the low-level functionality of SharpHook,
and a single class which implements all of them - this class actually
calls methods of the UioHook class.

The reason for these interfaces and this class is abstraction. Classes
such as SimpleReactiveGlobalHook provide a framework which builds upon
UioHook and adds a lot of functionality on top of it. But they are
tightly coupled to the static methods of that class. Providers add a
layer of abstraction to that. The high-level classes are now not
concerned about the low-level functionality - they don't care where it
comes from.

This enables 2 things primarily:

- Integration testing - mocking low-level classes is simpler than
  mocking high-level classes

- Substituting libuiohook with something else entirely while adapting
  to its interface without having to rewrite the high-level classes.
  This would make a theoretical implementation for Wayland on Linux
  much simpler for example.