Add click count when simulating mouse press/release
macOS requires that the current click count be specified when simulating
mouse press/release. E.g. when simulating a double click, the first
press/release pair will have the click count of 1, and the second pair
will have the click count of 2.

The global hook in libuiohook counts the clicks based on time, but
counting clicks the same way when simulating events is problematic so
the click count was added explicitly.

Click count is used only on macOS and is ignored on Windows and Linux