Add the old behaviour for test classes
Turns out that when test classes run with an event loop, it makes
testing real-world applications much more difficult and with no real
upside. Thus, the old behaviour was returned, with the ability to switch
between the simple behaviour and the event loop with the former being
the default one.